+++ backend/src/main/java/com/bungeobbang/backend/member/service/MemberService.java
@@ -3,7 +3,7 @@ package com.bungeobbang.backend.member.service;
 import com.bungeobbang.backend.common.exception.AuthException;
 import com.bungeobbang.backend.common.service.RedisService;
 import com.bungeobbang.backend.member.dto.response.LoginResponse;
-import com.bungeobbang.backend.member.infrastructure.JwtProvider;
+import com.bungeobbang.backend.common.infrastructure.JwtProvider;
 import com.bungeobbang.backend.member.domain.*;
 import com.bungeobbang.backend.member.domain.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
